+++ src/readline.h
@@ -39,6 +39,9 @@ extern jmp_buf g_readline_loc_info;
 extern bool    g_readline_loop;
 extern bool    g_resize_requested;
 
+extern bool g_hist_next;
+extern bool g_hist_prev;
+
 /*lint -sem(readline, r_null) */
 
 char	*readline           (const char *prompt);